# openapi.yaml
openapi: 3.0.1
info:
  title: Default module
  description: ''
  version: 1.0.0
...
paths:
...
  /tickets:
    get:
      summary: index
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: Accept
          in: header
          description: ''
          required: false
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples: {}
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              example:
                - id: 1
                  name: Technical Support
                  description: Hardware and software issues
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 2
                  name: Billing
                  description: Payment and invoicing inquiries
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 3
                  name: Feature Request
                  description: New feature suggestions
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 4
                  name: Bug Report
                  description: Software bugs and defects
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 5
                  name: General Inquiry
                  description: General questions and support
                  createdAt: '2026-08-26T15:33:58.356508'
          headers: {}
      security: []
    post:
      summary: store
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
        - name: X-User-Id
          in: header
          description: ''
          required: false
          example: '1'
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            example:
              category_id: 1
              title: test
              description: lorem
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties: {}
              example:
                - id: 1
                  name: Technical Support
                  description: Hardware and software issues
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 2
                  name: Billing
                  description: Payment and invoicing inquiries
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 3
                  name: Feature Request
                  description: New feature suggestions
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 4
                  name: Bug Report
                  description: Software bugs and defects
                  createdAt: '2026-08-26T15:33:58.356508'
                - id: 5
                  name: General Inquiry
                  description: General questions and support
                  createdAt: '2026-08-26T15:33:58.356508'
          headers: {}
      security: []
...
components:
...
  securitySchemes: {}
servers:
  - url: localhost:8080/api/
    description: Develop Env
security: []